feat: rename PhenotypicFeatures import and add new PhenotypicFeaturesView component
